feat: guiから複数配置
