+++ src/three/xyz-space/annotation/box-3d.ts
@@ -1,13 +1,14 @@
 import * as THREE from "three";
 
 import { get3dSpaceSceneInstance } from "../xyz-space";
-import { addAnnotationBoxToImageScene } from "../../rgb-image/annotated-box";
+import { addAnnotationBoxToImage } from "../../rgb-image/annotated-box";
 
 import { text } from "../../../html/element";
 
 // 箱を作成(デバッグ用)
 let boxId = 0;
-function createBox3d(x: number, y: number, z: number) {
+const annotationBoxes = new Array<THREE.Mesh>;
+function addAnnotationBoxTo3dSpace(x: number, y: number, z: number) {
     const geometry = new THREE.BoxGeometry(0.5, 0.5, 0.5);
     const material = new THREE.MeshBasicMaterial({ color: 0x00FFFF, wireframe: true });
     const box = new THREE.Mesh(geometry, material);
@@ -15,9 +16,19 @@ function createBox3d(x: number, y: number, z: number) {
     boxId++;
     box.position.set(x, y, z);
     get3dSpaceSceneInstance().add(box);
-    return box;
+    annotationBoxes.push(box)
+
+    function debug() {
+        let str = "";
+        for (let index = 0; index < annotationBoxes.length; index++) {
+            const annotationBox = annotationBoxes[index];
+            str += annotationBox.name + " : (" + annotationBox.position.x + ", " + annotationBox.position.y + ", " + annotationBox.position.z + ")\n";
+            text.value = str;
+        }
+    }
 }
 
+// TODO: クリックでアノテボックス配置
 document.addEventListener('mousedown', addAnnotationBox, false);
 function addAnnotationBox(event: MouseEvent) {
     // X座標
@@ -29,25 +40,15 @@ function addAnnotationBox(event: MouseEvent) {
     // console.log(event);
 }
 
-const annotationBoxes = new Array<THREE.Mesh>;
 function setBox() {
     // TODO: 右手座標系から左手座標系に変える
-    const annotationBox = createBox3d(box3d.x, box3d.y, box3d.z);
-    annotationBoxes.push(annotationBox)
-
-    addAnnotationBoxToImageScene(-box3d.x, box3d.y, box3d.z);
-
-    let str = "";
-    for (let index = 0; index < annotationBoxes.length; index++) {
-        const annotationBox = annotationBoxes[index];
-        str += annotationBox.name + " : (" + annotationBox.position.x + ", " + annotationBox.position.y + ", " + annotationBox.position.z + ")\n";
-        text.value = str;
-    }
+    addAnnotationBoxTo3dSpace(box3d.x_m, box3d.y_m, box3d.z_m);
+    addAnnotationBoxToImage(-box3d.x_m, box3d.y_m, box3d.z_m);
 }
 
 export const box3d = {
-    x: -1.6,
-    y: 1.8,
-    z: 6,
+    x_m: -1.6,
+    y_m: 1.8,
+    z_m: 6,
     set: function () { setBox() }
 }
